requirements.txt
torch>=2.0.0
torchvision>=0.15.0
datasets>=2.14.0
kaggle>=1.5.0
Pillow>=10.0.0
opencv-python>=4.8.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
editdistance>=0.6.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
tqdm>=4.65.0
